refactor : 쿠키 생성 기능 분리
- 주석 변경
- ResponseUtil.java 파일에 있던 쿠키 생성 부분을 CookieUtil 파일에 분리

#Issue : 251
