+++ module-api/src/main/java/com/checkping/api/controller/member/AuthController.java
@@ -1,5 +1,6 @@
 package com.checkping.api.controller.member;
 
+import com.checkping.api.auth.util.CookieUtil;
 import com.checkping.common.response.BaseResponse;
 import com.checkping.dto.member.request.LoginRequestDto;
 import com.checkping.service.member.auth.AuthService;
@@ -8,7 +9,10 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/")
@@ -20,24 +24,17 @@ public class AuthController implements AuthApi{
     /**
      * 로그인
      * - 성공 시 쿠키에 access, refresh 저장 → BaseResponse.success(...)
-     * - 실패 시 예외 발생 → GlobalExceptionHandler에서 처리
      */
     @PostMapping("/login")
     public BaseResponse<String> login(@RequestBody LoginRequestDto request, HttpServletResponse response) {
         // 1) Service 호출 (실패 시 예외 발생)
         AuthTokens tokens = authService.login(request.getEmail(), request.getPassword());
-
         // 2) 성공 시 쿠키 생성
-        Cookie accessCookie = new Cookie("access", tokens.getAccess());
-        accessCookie.setHttpOnly(true);
-        accessCookie.setPath("/");
+        Cookie accessCookie = CookieUtil.createCookie("access", tokens.getAccess());
         response.addCookie(accessCookie);
 
-        Cookie refreshCookie = new Cookie("refresh", tokens.getRefresh());
-        refreshCookie.setHttpOnly(true);
-        refreshCookie.setPath("/");
+        Cookie refreshCookie = CookieUtil.createCookie("refresh", tokens.getRefresh());
         response.addCookie(refreshCookie);
-
         // 3) 응답
         return BaseResponse.success("로그인에 성공하였습니다.");
     }
@@ -45,7 +42,6 @@ public class AuthController implements AuthApi{
     /**
      * 로그아웃
      * - 성공 시 쿠키 제거 → BaseResponse.success(...)
-     * - 실패 시 예외 발생 → GlobalExceptionHandler에서 처리
      */
     @PostMapping("/logout")
     public BaseResponse<String> logout(HttpServletRequest request, HttpServletResponse response) {
@@ -53,14 +49,10 @@ public class AuthController implements AuthApi{
         authService.logout(request);
 
         // 2) 쿠키 제거
-        Cookie delAccess = new Cookie("access", null);
-        delAccess.setMaxAge(0);
-        delAccess.setPath("/");
+        Cookie delAccess = CookieUtil.deleteCookie("access");
         response.addCookie(delAccess);
 
-        Cookie delRefresh = new Cookie("refresh", null);
-        delRefresh.setMaxAge(0);
-        delRefresh.setPath("/");
+        Cookie delRefresh = CookieUtil.deleteCookie("refresh");
         response.addCookie(delRefresh);
 
         // 3) 응답
